update
tùm lum tà la

+++ QLNhaTro_API/APIController/DICHVUController.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Web.Http;
 using QLNhaTro_API.Models;
@@ -10,61 +9,79 @@ namespace QLNhaTro_API.APIController
     {
         private DBQLNhaTro db = new DBQLNhaTro();
         // GET: api/DICHVU
-        public IEnumerable<DichVu> Get()
+        [HttpGet]
+        public IHttpActionResult Get()
         {
-            return db.DichVus.ToList();
+            return Ok(db.DichVus.ToList());
         }
 
         // GET: api/DICHVU/5
-        public DichVu Get(int id)
+        [HttpGet]
+        public IHttpActionResult Get(int id)
         {
-            return db.DichVus.SingleOrDefault(d => d.IdDichVu == id);
+            DichVu dichVu = db.DichVus.SingleOrDefault(p => p.IdDichVu == id);
+            if (dichVu == null)
+            {
+                return NotFound();
+            }
+            return Ok(dichVu);
         }
 
         // POST: api/DICHVU
-        public int Post(DichVu dichVu)
+        [HttpPost]
+        public IHttpActionResult Post(DichVu dichVu)
         {
             try
             {
                 if (!ModelState.IsValid)
                 {
-                    return 0;
+                    return Ok(new Message(0, "Thêm dịch vụ không thành công. Vui lòng thử lại"));
                 }
                 db.DichVus.Add(dichVu);
                 db.SaveChanges();
             }
             catch (Exception)
             {
-                return 0;
+                return Ok(new Message(0, "Thêm dịch vụ không thành công. Vui lòng thử lại"));
             }
-            return 1;
+            return Ok(new Message(1, "Thêm dịch vụ thành công"));
         }
 
         // PUT: api/DICHVU/5
-        public int Put(int id, DichVu newDichVu)
+        [HttpPut]
+        public IHttpActionResult Put(int id, DichVu newDichVu)
         {
             if (!ModelState.IsValid)
             {
-                return 0;
+                return Ok(new Message(0, "Thay đổi thông tin thất bại. Vui lòng kiểm tra và thử lại"));
             }
             var dichvu = db.DichVus.Find(id);
             if (dichvu == null)
             {
-                return -1;
+                return Ok(new Message(2, "Không tìm thấy dịch vụ cần thay đổi thông tin. Vui lòng kiểm tra và thử lại"));
             }
             dichvu.TenDichVu = newDichVu.TenDichVu;
             dichvu.DonGia = newDichVu.DonGia;
             db.SaveChanges();
-            return 1;
+            return Ok(new Message(1, "Thay đổi thông tin thành công"));
         }
 
         // DELETE: api/DICHVU/5
-        public int Delete(int id)
+        [HttpDelete]
+        public IHttpActionResult Delete(int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return Ok(new Message(0, "Xoá thất bại. Vui lòng kiểm tra và thử lại"));
+            }
             DichVu dichVu = db.DichVus.Find(id);
+            if (dichVu == null)
+            {
+                return Ok(new Message(2, "Không tìm thấy dịch vụ cần xoá. Vui lòng kiểm tra và thử lại"));
+            }
             db.DichVus.Remove(dichVu);
             db.SaveChanges();
-            return 1;
+            return Ok(new Message(1, "Xoá thành công"));
         }
     }
 }
